requirements.txt
pandas==2.2.3
scikit-learn==1.6.1
seaborn==0.13.2
streamlit==1.45.1
xgboost==3.0.2
numpy==2.2.6
matplotlib==3.10.3
streamlit-reveal-slides==0.2.0